moved stencils into stencils folder
