+++ model/atmosphere/dycore/src/icon4py/model/atmosphere/dycore/velocity_advection.py
@@ -16,12 +16,12 @@ from gt4py.next import backend as gtx_backend
 
 from icon4py.model.atmosphere.dycore import (
     compute_cell_diagnostics_for_velocity_advection,
-    compute_edge_diagnostics_for_velocity_advection,
     dycore_states,
 )
 from icon4py.model.atmosphere.dycore.stencils import (
     compute_advection_in_horizontal_momentum_equation,
     compute_advection_in_vertical_momentum_equation,
+    compute_edge_diagnostics_for_velocity_advection,
     compute_maximum_cfl_and_clip_contravariant_vertical_velocity,
 )
 from icon4py.model.common import dimension as dims, field_type_aliases as fa
