+++ model/atmosphere/dycore/tests/dycore_stencil_tests/test_compute_derived_horizontal_winds_and_ke_and_horizontal_advection_of_w_and_contravariant_correction.py
@@ -9,7 +9,7 @@ import gt4py.next as gtx
 import numpy as np
 import pytest
 
-from icon4py.model.atmosphere.dycore.compute_edge_diagnostics_for_velocity_advection import (
+from icon4py.model.atmosphere.dycore.stencils.compute_edge_diagnostics_for_velocity_advection import (
     compute_derived_horizontal_winds_and_ke_and_horizontal_advection_of_w_and_contravariant_correction,
 )
 from icon4py.model.common import dimension as dims
