+++ model/atmosphere/dycore/tests/dycore_stencil_tests/test_compute_horizontal_advection_of_w.py
@@ -9,7 +9,7 @@ import numpy as np
 import pytest
 
 import icon4py.model.testing.helpers as test_helpers
-from icon4py.model.atmosphere.dycore.compute_edge_diagnostics_for_velocity_advection import (
+from icon4py.model.atmosphere.dycore.stencils.compute_edge_diagnostics_for_velocity_advection import (
     compute_horizontal_advection_of_w,
 )
 from icon4py.model.common import dimension as dims
