+++ model/atmosphere/dycore/tests/dycore_tests/test_velocity_advection.py
@@ -12,13 +12,13 @@ import pytest
 
 from icon4py.model.atmosphere.dycore import (
     compute_cell_diagnostics_for_velocity_advection,
-    compute_edge_diagnostics_for_velocity_advection,
     dycore_states,
     velocity_advection as advection,
 )
 from icon4py.model.atmosphere.dycore.stencils import (
     compute_advection_in_horizontal_momentum_equation,
     compute_advection_in_vertical_momentum_equation,
+    compute_edge_diagnostics_for_velocity_advection,
 )
 from icon4py.model.common import dimension as dims, utils as common_utils
 from icon4py.model.common.grid import (
